Added Pundit for user authorization, updated ApplicationPolicy and added WikiPolicy, and updated views so only authorized users can create, edit, and delete wikis. Currently, any authenticated user can edit a wiki, but only the creator can destroy.

+++ app/controllers/wikis_controller.rb
@@ -1,18 +1,22 @@
 class WikisController < ApplicationController
   def index
     @wikis = Wiki.all
+    authorize @wikis
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
     def create
       @wiki = current_user.wikis.build(params.require(:wiki).permit(:title, :body, :private))
+      authorize @wiki
 
       if @wiki.save
         flash[:notice] = "#{@wiki.title} saved!"
@@ -25,10 +29,12 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
     def update
       @wiki = Wiki.find(params[:id])
+      authorize @wiki
       
       if @wiki.update_attributes(params.require(:wiki).permit(:title, :body, :private))
         flash[:notice] = "#{@wiki.title} saved!"
@@ -41,12 +47,14 @@ class WikisController < ApplicationController
 
     def destroy
       @wiki = Wiki.find(params[:id])
+      authorize @wiki
 
       if @wiki.destroy
         flash[:notice] = "#{@wiki.name} has been deleted."
         redirect_to wikis_path
       else
         flash[:error] = "There was a problem deleting your article. Please try again."
+        render :edit
       end
     end
 end
